修复队列长度不能大于255的bug：数据类型错误。

+++ MasterNet/MasterNet/MasterNet/yy-file/src/FifoQueue/fifoqueue.h
@@ -21,7 +21,7 @@
 #pragma pack(push, 1)
 typedef struct{
 	char		data[512];
-	uint8_t		len;
+	int			len;
 
 }fifoqueue_t;
